% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{absolute_error}
\alias{absolute_error}
\title{Absolute error}
\usage{
absolute_error(model, df_data = NA)
}
\arguments{
\item{model}{sBG model}

\item{df_data}{data frame that contains per cohort and period the survived customers (must have 
columns 'cohort', 'period' and 'survived_customers').}
}
\value{
a matrix with absolute errors per cohort (rows) and period (columns)
}
\description{
Computes absolute error (actual retention rate - predicted retention rate) per cohort and period 
for given input data. Input data can either be used from the model (only available when trained on
data) or provided as a parameter.
}
