% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer_performance.R
\name{customer_rates}
\alias{customer_rates}
\title{Customer performance rates}
\usage{
customer_rates(model, start_period, end_period, metrics = c("churn",
  "retention", "survival"))
}
\arguments{
\item{model}{a sBG model}

\item{start_period}{a number denoting the period to start calculation from}

\item{end_period}{a number denoting the period to end calculation at}

\item{metrics}{vector with customer performance rate to compute.
Default is a vector contain all three rates ('churn', 'retention' and 'survival').}
}
\value{
a data frame with column period and given customer performance rate
}
\description{
\code{customer_rates} calculates churn, retention and survival rate
using the parameters from model input for periods between start_period and end_period.
}
