% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer_performance.R
\name{derl}
\alias{derl}
\title{Calculate discounted expected residual lifetime}
\usage{
derl(model, period = 1, discount_rate = 0.25)
}
\arguments{
\item{model}{a sBG model}

\item{period}{a vector with period(s) for which DERL should be calculated. (customer has made period-1 renewals). Default is set to 1.}

\item{discount_rate}{a number between 0 and 1 denoting the discount rate. Default is set to 0.25 (recommended value for private businesses).}
}
\value{
a vector with discounted expected residual lifetime per given period
}
\description{
\code{derl} computes discounted expected residual lifetime (DERL) for given customer's contract period(s)
}
\references{
http://brucehardie.com/papers/022/fader_hardie_mksc_10.pdf Equation (6)
}
