% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sBGmodel.R
\name{fit}
\alias{fit}
\title{Fit sBG Model}
\usage{
fit(df_data, cp_period = 0)
}
\arguments{
\item{df_data}{a data frame must contain the columns 'cohort', 'period' and 'survived_customers'.}

\item{cp_period}{a number which denotes the period after which any change in the behaviour is happening.
If it is equal to 0, no change is happening. Default is set as 0.
Example: If period one shows a different behaviour as the other periods, cp_period is set as 1.}
}
\value{
list with the following components: 
'params' a list with estimated parameter, 
'df_data' the data frame used for fitting.
}
\description{
\code{fit} fits a sBG model for CLV. It estimates the parameter \code{alpha}, \code{beta}, \code{cp_alpha} and \code{cp_beta}
The model is evaluated by calculating the median and mean of the root mean-square error between the predicted and actual retention rate per cohort.
}
