% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sBGmodel.R
\name{multi_cohort_ll}
\alias{multi_cohort_ll}
\title{Calculate log-likelihood for multi-cohort data}
\usage{
multi_cohort_ll(m_data, params)
}
\arguments{
\item{m_data}{a lower triangular matrix containing number of customers per period and cohort. Needs to be squared with more than 1 column.}

\item{params}{a list of parameters \code{alpha}, \code{beta} to calculate log-likelihood function for.}
}
\value{
Log-likelihood based on multiple cohort data and given input parameters.
}
\description{
\code{multi_cohort_ll} computes log-likelihood function for a multi-cohort data. Input matrix must be a full lower triangular matrix.
}
\references{
http://brucehardie.com/notes/017/sBG_estimation.pdf: Equation (3)
}
