% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{performance_metrics}
\alias{performance_metrics}
\title{Performace metrics}
\usage{
performance_metrics(model, df_data = NA, metrics = "absolute_error")
}
\arguments{
\item{model}{sBG model}

\item{df_data}{data frame that contains per cohort and period the survived customers (must have columns 
'cohort', 'period' and 'survived_customers').}

\item{metrics}{list of performance metrics names you want to get (currently only absolute_error is available).}
}
\value{
a vector with absolute error per period (period 1 is first element).
}
\description{
\code{performance_metrics} computes performance metrics for sBG model fit. Input data can 
either be used from the model (only available when trained on data) or provided as a parameter.
}
