% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer_performance.R
\name{retention_rates}
\alias{retention_rates}
\title{Calculate retention rate}
\usage{
retention_rates(params, period)
}
\arguments{
\item{params}{a list of model parameters \code{alpha}, \code{beta}, \code{cp_alpha}, \code{cp_beta} and \code{cp_period}.
If it only contains \code{alpha} and \code{beta}, \code{cp_alpha} and \code{cp_beta} are set to \code{NA} and \code{cp_period} is set to 0.}

\item{period}{a numeric vector. Denotes for for which period(s) to calculate the retention probabilities.}
}
\value{
a vector with retention probabilities per period(s)
}
\description{
\code{retention_rates} computes retention probabilities for given periods based on the sBG model.
}
\references{
http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.558.935&rep=rep1&type=pdf: Equation (8)
}
