% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sBGmodel.R
\name{sBG_model}
\alias{sBG_model}
\title{Initialize sBG Model}
\usage{
sBG_model(alpha, beta, cp_alpha = NA, cp_beta = NA, cp_period = 0)
}
\arguments{
\item{alpha}{alpha parameter for sBG model}

\item{beta}{beta parameter for sBG model}

\item{cp_alpha}{alpha parameter for sBG model for periods until changepoint (when cp_period > 0)}

\item{cp_beta}{beta parameter for sBG model for periods until changepoint (when cp_period > 0)}

\item{cp_period}{changepoint period}
}
\value{
list with the following components: 'params' a list with given parameter
}
\description{
Initialize a sBG model with parameter list
}
