% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sBGmodel.R
\name{sBG_parameters}
\alias{sBG_parameters}
\title{Parameter list for sBG model.}
\usage{
sBG_parameters(alpha, beta, cp_alpha = NA, cp_beta = NA,
  cp_period = 0)
}
\arguments{
\item{alpha}{alpha paramter for sBG model}

\item{beta}{beta paramter for sBG model}

\item{cp_alpha}{alpha paramter for sBG model for periods until changepoint (when cp_period > 0)}

\item{cp_beta}{beta paramter for sBG model for periods until changepoint (when cp_period > 0)}

\item{cp_period}{changepoint period (default is 0, which means no change point)}
}
\value{
list with parameters for sBG model
}
\description{
\code{sBG_parameters} takes the parameter in the argument and combines them to one list as it is also in the output of the model
}
