% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sBGmodel.R
\name{single_cohort_data}
\alias{single_cohort_data}
\title{Transform input for single cohort sBG model}
\usage{
single_cohort_data(df_data)
}
\arguments{
\item{df_data}{a data frame that contains per cohort and period the survived customers (must have columns 'cohort', 'period' and 'survived_customers').}
}
\value{
a data frame that contains per cohort and period the survived and lost customers (excluding first period).
}
\description{
\code{single_cohort_data} creates a data frame which can be used to apply the sBG model for a single cohort.
}
