% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sBGmodel.R
\name{single_cohort_ll}
\alias{single_cohort_ll}
\title{Log-likelihood for single-cohort data}
\usage{
single_cohort_ll(df_single_cohort_data, params)
}
\arguments{
\item{df_single_cohort_data}{a data.frame containing number of lost and survived_customers per period. 
Required columns:  columns 'period', 'lost_customers' and 'survived_customers'.}

\item{params}{a list of parameters \code{alpha}, \code{beta} to calculate log-likelihood function for.}
}
\value{
Log-likelihood based on single cohort data and given input parameters.
}
\description{
\code{single_cohort_ll} computes log-likelihood function for single-cohort data.
}
\references{
http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.558.935&rep=rep1&type=pdf, formula B3.
}
