% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sBGmodel.R
\name{single_point_parameters}
\alias{single_point_parameters}
\title{sBG model parameters}
\usage{
single_point_parameters(var, mu)
}
\arguments{
\item{var}{variance}

\item{mu}{mean}
}
\value{
list with derived parameters for sBG model
}
\description{
\code{single_point_parameters} estimates parameters for sBG model based on variance and mean.
}
