% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer_performance.R
\name{survival_rates}
\alias{survival_rates}
\title{Calculate survival rate}
\usage{
survival_rates(params, period)
}
\arguments{
\item{params}{a list of model parameters \code{alpha}, \code{beta}, \code{cp_alpha}, \code{cp_beta} and \code{cp_period}.
If it only contains \code{alpha} and \code{beta}, \code{cp_alpha} and \code{cp_beta} are set to \code{NA} and \code{cp_period} is set to 0.}

\item{period}{a numeric vector. Denotes for for which period(s) to calculate the survival probabilities.}
}
\value{
a vector with survival probabilities per period(s)
}
\description{
\code{survival_rates} computes survival probabilities for given periods based on sBG model.
}
\references{
http://brucehardie.com/notes/017/sBG_estimation.pdf: Equation (2)
}
