% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sBGmodel.R
\name{triangular_matrix}
\alias{triangular_matrix}
\title{Transform dataframe to triangular matrix for CLV model.}
\usage{
triangular_matrix(df_data)
}
\arguments{
\item{df_data}{a data frame that contains per cohort and period the survived customers (must have columns 'cohort', 'period' and 'survived_customers').}
}
\value{
a lower triangular matrix containing number of customers per cohort and period
}
\description{
\code{triangular_matrix} computes lower triangular matrix for the CLV model.
}
