% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{validate_input_data}
\alias{validate_input_data}
\title{Validate input data for sBG model}
\usage{
validate_input_data(df_data)
}
\arguments{
\item{df_data}{a data frame must contain the columns 'cohort', 'period' and 'survived_customers'.}
}
\description{
\code{validate_input_data} receives a dataframe and validated if all columns are given and no NA values are included.
}
