% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{validate_input_model}
\alias{validate_input_model}
\title{Validate input of sBG model}
\usage{
validate_input_model(model)
}
\arguments{
\item{model}{sBG model.}
}
\description{
\code{validate_input_model} receives a model and validates if its structure is correct (i.e. if it contains the right components).
}
