% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{validate_input_params}
\alias{validate_input_params}
\title{Validate input parameter for sBG model}
\usage{
validate_input_params(params)
}
\arguments{
\item{params}{a list of parameters (\code{alpha}, \code{beta}, \code{cp_alpha}, \code{cp_beta} and \code{cp_period}).}
}
\description{
\code{validate_input_params} receives list with parameters and validates if all of them are correctly given.
}
